requirements.txt
pyserial>=3.5
numpy>=1.24
matplotlib>=3.7
neurokit2>=0.2